requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
fastapi==0.75.0
frozenlist==1.3.0
gql==3.1.0
graphql-core==3.2.0
h11==0.12.0
httpcore==0.14.7
idna==3.3
multidict==6.0.2
pydantic==1.9.0
Pygments==2.11.2
python-dotenv==0.19.2
rfc3986==1.5.0
rich==10.16.2
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
yarl==1.7.2
